First system test: Sit And Play
One of the significant features of the game is the "Sit And Play"
system: a new player can join at any time during the game and
start playing. Since this is both a important feature and a good
use case for Hotwire (because it implies to update the HTML
of all players in real time), it makes for a good first feature
to implement.

In true BDD fashion, we'll let our test guide our implementation,
including the UI – even though we also have one mockup or two in
reserve.

This first system test is quite simple: we open a session in the
middle of a game, join this game in another session, and assert
that the players list is up-to-date in both sessions – i.e. that
the view in the first session has been updated.

This test reveals several pieces of business logic that we now
have to implement: a `Game` model, some kind of phase counter for
said model, and a list of players.