First resources: game and players
Our system test fails at the very beginning, because we lack the resources
that it mentions: a game and players participating in this game.

These are pretty straightforward and easy to generate. All we know so far
about the `Game` model is that it has a _call_, that is somehow numeric.
(In truth, we can guess that it will certainly be some kind of `Enum`,
but this hasn't been revealed by our test yet).

The `Player` model is even simpler: it represents a participant in a
given game. At the moment, a player is defined by very few attributes:
a name and a position at the table – this will be significant later on,
since it's the relative position of a player to another that determines
the attacks that can be made against them.