Experiment: extract the players list to a model
Instead of being a simple data collection provided by the Game model,
the list of players ordered for a given player is now a self-contained
model.

On one hand, this allows the leverage of Rails' conventions (something
visible in the simplification of the call to
`#broadcast_replace_later_to`), and technically enforces the Single
Responsibility Principle, by reducing the scope of the Game model.

On the other hand, the end result is a bit clunky – in particular the
way the `PlayersList` model has to implement `#id` and `.find` so
that it can be serialized to a GlobalID and passed to ActiveJob. And the
business logic behind this list is so simple that extracting it from
the `Game` model doesn't seem that important.

I want to keep this approach for future reference, but not in the
definitive codebase – this commit will be reverted immediately.