Failed attempt at a new feature: quitting a game
When a user closes their browser window, the websocket connection is closed;
which means than we should be able to update accordingly, since
ActionCable has a callback method for this event.

However, this requires to override the `#unsubscribe` method of the channel,
and Turbo doesn't let us use a custom channel – not without losing most
of its sugar and magic anyway. So this will not work.

Let's try something else instead: leaving by consuming a specific RESTful
resource.